Allow for mono feature functions

+++ tests/test_process.py
@@ -637,7 +637,7 @@ def test_process_signal_from_index(
     if not expected:
         pd.testing.assert_series_equal(
             result,
-            pd.Series([], index),
+            pd.Series([], index, dtype=float),
         )
     else:
         pd.testing.assert_series_equal(
